requirements.txt
numpy
pandas
azure-cognitiveservices-language-luis==0.7.0
sklearn